Restructure the add cluster flow
To improve the experience with kubenav, we changed the flow on how to
add clusters. The add cluster modal only shows the available options
now. When a option is selected the user is redirected to the
corresponding page to add a cluster.
